improved clear command, limit is 50msg

+++ src/commands/moderation/clearChat.ts
@@ -1,4 +1,4 @@
-import { Message } from 'discord.js';
+import { Message, DiscordAPIError } from 'discord.js';
 import { tsukasaConfig, ServerConfig } from '../../main';
 import { readFileSync } from 'fs';
 import permissionCheck from '../../utility/permissionCheck';
@@ -25,9 +25,19 @@ const clearChat = async (args: string[], msg: Message) => {
         return;
     }
 
-    await msg.channel.bulkDelete(stringToNumber);
-    await msg.channel.send("I cleared " + stringToNumber + " messages :heart:");
-    msg.delete(5000);
+    if (!(stringToNumber < 50)) {
+        msg.reply("don't delete more then 50 messages please, thanks :D :heart:");
+        return;
+    }
 
+    try {
+
+        await msg.channel.bulkDelete(stringToNumber);
+        await msg.channel.send("I cleared " + stringToNumber + " messages :heart:");
+
+    } catch (error) {
+        console.log(error);
+        msg.reply("an error occurred please note you can't delete messages that are older then 14 days");
+    }
 }
 export default clearChat;
